deletion fakeExpense

+++ app/src/main/java/com/loconav/locodriver/expense/model/AddExpenseRequestBody.kt
@@ -7,7 +7,7 @@ data class AddExpenseRequestBody(
     @SerializedName("expense_type")
     var expenseType: String? = null,
     @SerializedName("amount")
-    var amount: Int? = null,
+    var amount: Double? = null,
     @SerializedName("expense_date")
     var expenseDate: Long? = null,
     @SerializedName("remarks")
